Type of PI should be float not double due to ImGuiAPI

+++ ImPlotGraphs/hdr/ImPlotGraph3.hpp
@@ -9,7 +9,7 @@
 namespace Code::ImGuiImPlot
 {
     
-extern const double PI;
+extern const float PI;
 
 class Plot3 : public PlotGeneric
 {
